Make a separate package for reports
